requirements_test.txt
pytest==3.0.7
pytest-cov==2.4.0
pytest-xdist==1.16.0
requests-mock==1.3.0